Use a more compact JSON with verbs generated from wordnet
The wordnet package is large and its db folder is 28mb in size.
Because this package will be used in a GitHub action, we want the clone process to be as fast as possible.
Having that said, this commit adds a script that generates a very small <= 200KB file with just the English verbs.